+++ src/Firebase/Auth/ApiClient.php
@@ -12,7 +12,6 @@ use Kreait\Firebase\Exception\Auth\OperationNotAllowed;
 use Kreait\Firebase\Exception\Auth\UserDisabled;
 use Kreait\Firebase\Exception\AuthApiExceptionConverter;
 use Kreait\Firebase\Exception\AuthException;
-use Kreait\Firebase\Exception\FirebaseException;
 use Kreait\Firebase\Request;
 use Kreait\Firebase\Util\JSON;
 use Kreait\Firebase\Value\Provider;
@@ -41,7 +40,6 @@ class ApiClient
 
     /**
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function createUser(Request\CreateUser $request): ResponseInterface
     {
@@ -50,7 +48,6 @@ class ApiClient
 
     /**
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function updateUser(Request\UpdateUser $request): ResponseInterface
     {
@@ -61,7 +58,6 @@ class ApiClient
      * @param array<string, mixed> $claims
      *
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function setCustomUserClaims(string $uid, array $claims): ResponseInterface
     {
@@ -76,7 +72,6 @@ class ApiClient
      *
      * @throws EmailNotFound
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function getUserByEmail(string $email): ResponseInterface
     {
@@ -87,7 +82,6 @@ class ApiClient
 
     /**
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function getUserByPhoneNumber(string $phoneNumber): ResponseInterface
     {
@@ -98,7 +92,6 @@ class ApiClient
 
     /**
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function downloadAccount(?int $batchSize = null, ?string $nextPageToken = null): ResponseInterface
     {
@@ -112,7 +105,6 @@ class ApiClient
 
     /**
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function deleteUser(string $uid): ResponseInterface
     {
@@ -122,10 +114,31 @@ class ApiClient
     }
 
     /**
+     * @param string[] $uids
+     *
+     * @throws AuthException
+     */
+    public function deleteUsers(string $projectId, array $uids, bool $forceDeleteEnabledUsers, ?string $tenantId = null): ResponseInterface
+    {
+        $data = [
+            'localIds' => $uids,
+            'force' => $forceDeleteEnabledUsers,
+        ];
+
+        if ($tenantId) {
+            $data['tenantId'] = $tenantId;
+        }
+
+        return $this->requestApi(
+            "https://identitytoolkit.googleapis.com/v1/projects/{$projectId}/accounts:batchDelete",
+            $data
+        );
+    }
+
+    /**
      * @param string|array<string> $uids
      *
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function getAccountInfo($uids): ResponseInterface
     {
@@ -143,7 +156,6 @@ class ApiClient
      * @throws InvalidOobCode
      * @throws OperationNotAllowed
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function verifyPasswordResetCode(string $oobCode): ResponseInterface
     {
@@ -158,7 +170,6 @@ class ApiClient
      * @throws OperationNotAllowed
      * @throws UserDisabled
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function confirmPasswordReset(string $oobCode, string $newPassword): ResponseInterface
     {
@@ -170,7 +181,6 @@ class ApiClient
 
     /**
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function revokeRefreshTokens(string $uid): ResponseInterface
     {
@@ -184,7 +194,6 @@ class ApiClient
      * @param array<int, string|Provider> $providers
      *
      * @throws AuthException
-     * @throws FirebaseException
      */
     public function unlinkProvider(string $uid, array $providers): ResponseInterface
     {
@@ -198,14 +207,15 @@ class ApiClient
      * @param array<mixed> $data
      *
      * @throws AuthException
-     * @throws FirebaseException
      */
     private function requestApi(string $uri, array $data): ResponseInterface
     {
         $options = [];
+        $tenantId = $data['tenantId'] ?? $this->tenantId ?? null;
+        $tenantId = $tenantId instanceof TenantId ? $tenantId->toString() : $tenantId;
 
-        if ($this->tenantId) {
-            $data['tenantId'] = $this->tenantId->toString();
+        if ($tenantId) {
+            $data['tenantId'] = $tenantId;
         }
 
         if (!empty($data)) {